requirements.txt
aiohttp==3.8.1
flask-sqlalchemy==3.0.3
flask==2.3.2
celery==5.3.1
beautifulsoup4==4.11.1
pymysql==1.0.3
redis==4.5.5
requests==2.28.0